Move search-bar component into its own directory
