+++ frontend/src/js/Header/SearchBar/search-icon.png
